% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_easy_calc.R
\name{regr_easy_calc}
\alias{regr_easy_calc}
\title{Calculation of Regression Models: Linear, Quadratic and Cubic.}
\usage{
regr_easy_calc(x, y, model = "all")
}
\arguments{
\item{x}{Values that should be used as an independent variable for the regression calculation.}

\item{y}{Values that should be used as a dependent variable for the regression calculation.}

\item{model}{Character, defined which model will be calculated. model = "L", calculate the linear, model = "Q" calculate the quadratic, model = "C" calculate the cubic, model = "all" = calculate both).}
}
\value{
returns a list with the regression result (linear, quadratic and/or cube)
}
\description{
Performs regression calculations: linear, quadratic and cubic, allowing to perform only one or both, returning a detailed result of the calculation
}
\examples{
library(regr.easy)
x <- seq(0,300,50)
y <- c(138.6,153.6,164.525,164.925,158.725,159.975,154.425)
regr_easy_calc(x,y,model = "all")
}
